requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.1
django-appconf==1.0.5
django-cryptography==1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
psycopg-binary==3.0.16
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
